// apps/next/lib/deploy/gen/argoproj.io.ts
export class Rollout extends ApiObject {
  /**
   * Returns the apiVersion and kind for "Rollout"
   */
  public static readonly GVK: GroupVersionKind = {
    apiVersion: 'argoproj.io/v1alpha1',
    kind: 'Rollout',
  }

  /**
   * Renders a Kubernetes manifest for "Rollout".
   *
   * This can be used to inline resource manifests inside other objects (e.g. as templates).
   *
   * @param props initialization props
   */
  public static manifest(props: RolloutProps): any {
    return {
      ...Rollout.GVK,
      ...toJson_RolloutProps(props),
    };
  }

  /**
   * Defines a "Rollout" API object
   * @param scope the scope in which to define this object
   * @param id a scope-local name for the object
   * @param props initialization props
   */
  public constructor(scope: Construct, id: string, props: RolloutProps) {
    super(scope, id, {
      ...Rollout.GVK,
      ...props,
    });
  }

  /**
   * Renders the object to Kubernetes JSON.
   */
  public toJson(): any {
    const resolved = super.toJson();

    return {
      ...Rollout.GVK,
      ...toJson_RolloutProps(resolved),
    };
  }
}

/**
 * @schema Rollout
 */
export interface RolloutProps {
  /**
   * @schema Rollout#metadata
   */
  readonly metadata?: ApiObjectMetadata;

  /**
   * @schema Rollout#spec
   */
  readonly spec: RolloutSpec;

}
// ...
export interface RolloutSpec {
  /**
   * @schema RolloutSpec#analysis
   */
  readonly analysis?: RolloutSpecAnalysis;

  /**
   * @schema RolloutSpec#minReadySeconds
   */
  readonly minReadySeconds?: number;

  /**
   * @schema RolloutSpec#paused
   */
  readonly paused?: boolean;

  /**
   * @schema RolloutSpec#progressDeadlineAbort
   */
  readonly progressDeadlineAbort?: boolean;

  /**
   * @schema RolloutSpec#progressDeadlineSeconds
   */
  readonly progressDeadlineSeconds?: number;

  /**
   * @schema RolloutSpec#replicas
   */
  readonly replicas?: number;

  /**
   * @schema RolloutSpec#restartAt
   */
  readonly restartAt?: Date;

  /**
   * @schema RolloutSpec#revisionHistoryLimit
   */
  readonly revisionHistoryLimit?: number;

  /**
   * @schema RolloutSpec#rollbackWindow
   */
  readonly rollbackWindow?: RolloutSpecRollbackWindow;

  /**
   * @schema RolloutSpec#selector
   */
  readonly selector?: RolloutSpecSelector;

  /**
   * @schema RolloutSpec#strategy
   */
  readonly strategy?: RolloutSpecStrategy;

  /**
   * @schema RolloutSpec#template
   */
  readonly template?: RolloutSpecTemplate;

  /**
   * @schema RolloutSpec#workloadRef
   */
  readonly workloadRef?: RolloutSpecWorkloadRef;

}